Disentangle Simulation interface and delete MEWS
